Change the shape used in the uncrouch collision check from box to sphere
This change ensures the collision volume used to check whether the
player can uncrouch matches the actual collision volume of the
player character, geometrically. Previously, a box collider was
used for the check which would lead to mismatches which would
prevent the player character from uncrouching in places they'd
expect to be able to, e.g. in corners.