Add tests to validate that doors can only lock when closed
Adds two new tests; Test_Door_OpenDoorCannotLock and
Test_Door_ClosedDoorCanLock to validate that doors can only lock when
closed. Also made the IsLocked and IsOpen variables in BP_Door
private and changeable only through getters and setters. This allows
the functional tests to more easily set up their state, as the setters
can make sure everything about the door is in the correct state when
the state is changed.