Add a death camera when the player dies
Adds the BP_DeathCam actor which spawns when the player dies.
The DeathCam is just a camera attached to a sphere which spawns
at the position of the first-person camera on death and falls to
the ground. The physics properties of the sphere are such that it
rolls to one side and stays still.
This change also updates the game so the player dies on contact with
the Securibot.