Add blueprint interface and enum to support glowing outlines
The BPI_OutlineGlow blueprint interface should be implemented by
any actor to which a glowing outline should be applied. The
E_PostProcessingOutlineChannel enum is used in tandem with the
interface and allows us to define different colours/channels for
the outline in different circumstances. The actual glowing outline
will be applied by the FirstpersonCharacter blueprint, but the
details of the outline will be handled by the actor which implements
BPI_OutlineGlow