Update FirstPersonCharacter blueprint to constantly apply glow effects
This change updates BP_FirstPersonCharacter with a new event which
runs on a 0.1 second loop. Each time the event runs, it performs a
raycast using the same vectors as would be used to interact with
objects. If anything is detected within range, we will attempt to
call the SetOutlineGlow event on the target object. This event will
only be present if the target object implements  BPI_OutlineGlow.
The loop will also check if we've stopped looking at an object and
disable the glow if so.