Update Securibot AI to investigate Eliza noises
Update the Securibot behaviour tree such that if it hears an Eliza
chime while on patrol, it will move to investigate the source of the
noise. Once the noise is investigated, it returns to its patrol path.
If the Securibot senses something more interesting, like the player,
during its investigation, it will abandon the Eliza noise and pursue
the player.