Update doors so they always open away from the player
Not strictly realistic, but this change updates dooes so they will
always swing away from the player. This is a convention I've seen
in a number of games and it makes gameplay smoother, as players will
typically walk very close to a door to interact with it. If the
door then swings back in their face it can push the player character
around and cause a jarring bump as the player collision capsule moves.