Highlight all interactable objects in player's view
To make interactable objects glow, this commit adds a new collision
channel called ARScan. The player character blueprint performs a box
trace every 0.1 seconds to find objects in the player's FOV which
implement BP_interactable but are not close enough to interact with
or are not the currently selected actor.