Update the way line of sight is detected for objectives
The previous method for detecting line of sight with an objective was
flaky because it only checked line of sight to the midpoint of the
object. After this change, 3 key points on the mesh are checked and
if the player has line of sight to any of them, the check succeeds.