Update Securibot warning so it works when Securibot cannot see player
The boolean controlling Securibot's behaviour while issuing warnings
to the player was only updated while it could see the player. This could
result in the bot getting stuck if the player left the area while being
warned. This commit changes behaviour so the update is independent of
the bot being able to see the player.